Added Python 3 & Django 1.6 support
