+++ .gitignore
@@ -1,3 +1,5 @@
 .DS_Store
 *.pyc
+.idea
+*.egg-info
 *~
