+++ .travis.yml
@@ -3,10 +3,13 @@ python:
   - "2.5"
   - "2.6"
   - "2.7"
+  - "3.2"
+  - "3.3"
 before_install:
   - export PIP_INDEX_URL=https://simple.crate.io/
 install:
   - pip install Django==$DJANGO_VERSION
+  - pip install six
   - pip install -e .
 script:
   - make test
@@ -14,8 +17,9 @@ env:
   - DJANGO_VERSION=1.1.4
   - DJANGO_VERSION=1.2.7
   - DJANGO_VERSION=1.3.7
-  - DJANGO_VERSION=1.4.5
-  - DJANGO_VERSION=1.5
+  - DJANGO_VERSION=1.4.9
+  - DJANGO_VERSION=1.5.5
+  - DJANGO_VERSION=https://github.com/django/django/zipball/stable/1.6.x
 branches:
   only:
     - master
@@ -23,3 +27,21 @@ matrix:
   exclude:
     - python: "2.5"
       env: DJANGO_VERSION=1.5
+    - python: "2.5"
+      env: DJANGO_VERSION=https://github.com/django/django/zipball/stable/1.6.x
+    - python: "3.2"
+      env: DJANGO_VERSION=1.1.4
+    - python: "3.2"
+      env: DJANGO_VERSION=1.2.7
+    - python: "3.2"
+      env: DJANGO_VERSION=1.3.7
+    - python: "3.2"
+      env: DJANGO_VERSION=1.4.9
+    - python: "3.3"
+      env: DJANGO_VERSION=1.1.4
+    - python: "3.3"
+      env: DJANGO_VERSION=1.2.7
+    - python: "3.3"
+      env: DJANGO_VERSION=1.3.7
+    - python: "3.3"
+      env: DJANGO_VERSION=1.4.9
