+++ treemenus/templatetags/tree_menu_tags.py
@@ -2,6 +2,10 @@ import django
 from django import template
 from django.template.defaulttags import url
 from django.template import Node, TemplateSyntaxError
+try:
+    from django.utils import six
+except ImportError:  # Django < 1.5
+    import six
 
 from treemenus.models import Menu, MenuItem
 from treemenus.config import APP_LABEL
@@ -32,7 +36,8 @@ register.inclusion_tag('%s/menu.html' % APP_LABEL, takes_context=True)(show_menu
 
 def show_menu_item(context, menu_item):
     if not isinstance(menu_item, MenuItem):
-        raise template.TemplateSyntaxError, 'Given argument must be a MenuItem object.'
+        six.reraise(template.TemplateSyntaxError, 'Given argument must be a '
+                                                  'MenuItem object.')
 
     context['menu_item'] = menu_item
     return context
@@ -49,9 +54,9 @@ class ReverseNamedURLNode(Node):
 
         resolved_named_url = self.named_url.resolve(context)
         if django.VERSION >= (1, 3):
-            contents = u'url "%s"' % resolved_named_url
+            contents = 'url "%s"' % resolved_named_url
         else:
-            contents = u'url %s' % resolved_named_url
+            contents = 'url %s' % resolved_named_url
 
         urlNode = url(self.parser, Token(token_type=TOKEN_BLOCK, contents=contents))
         return urlNode.render(context)
