+++ treemenus/tests/fake_menu_extension/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
         
-from models import FakeMenuItemExtension
+from .models import FakeMenuItemExtension
         
 admin.site.register(FakeMenuItemExtension)
